app/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
prometheus_client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-instrumentation-asgi==0.40b0
sqlalchemy==2.0.23
opentelemetry-instrumentation-sqlalchemy==0.40b0
opentelemetry-instrumentation-requests==0.40b0
requests==2.31.0
pydantic[email]==1.10.8
